requirements.txt
numpy
matplotlib
pandas
streamlit
TA-Lib
requests
python-dateutil
